set all the hot start setup in unbuiltsolver directly
