+++ Sofa/Component/Constraint/Lagrangian/Solver/src/sofa/component/constraint/lagrangian/solver/GenericConstraintSolver.h
@@ -27,7 +27,6 @@
 #include <sofa/core/behavior/BaseConstraintCorrection.h>
 #include <sofa/core/behavior/BaseLagrangianConstraint.h>
 #include <sofa/helper/map.h>
-#include <sofa/simulation/mechanicalvisitor/MechanicalGetConstraintInfoVisitor.h>
 
 #include <sofa/simulation/task/CpuTask.h>
 #include <sofa/helper/OptionsGroup.h>
@@ -68,7 +67,6 @@ public:
     Data<bool> d_scaleTolerance; ///< Scale the error tolerance with the number of constraints
     Data<bool> d_allVerified; ///< All constraints must be verified (each constraint's error < tolerance)
 
-    Data<bool> d_initialGuess; ///< Activate constraint force history to improve convergence (hot start)
     Data<bool> d_computeGraphs; ///< Compute graphs of errors and forces during resolution
     Data<std::map < std::string, sofa::type::vector<SReal> > > d_graphErrors; ///< Sum of the constraints' errors at each iteration
     Data<std::map < std::string, sofa::type::vector<SReal> > > d_graphConstraints; ///< Graph of each constraint's error at the end of the resolution
@@ -98,6 +96,9 @@ protected:
     sofa::core::MultiVecDerivId m_dxId;
 
     virtual void initializeConstraintProblems();
+    virtual void preApplyCorrection() { doPreApplyCorrection(); }
+    void preClearCorrection(const core::ConstraintParams* cparams) { doPreClearCorrection(cparams); }
+    void postClearCorrection() { doPostClearCorrection(); }
 
     /*****
      *
@@ -124,37 +125,12 @@ protected:
      *
      */
     virtual void doSolve( GenericConstraintProblem * problem, SReal timeout = 0.0) = 0;
+    
+    virtual void doPreApplyCorrection() { }
+    virtual void doPreClearCorrection(const core::ConstraintParams* cparams) { SOFA_UNUSED(cparams); }
+    virtual void doPostClearCorrection() { }
 
     virtual void addRegularization(linearalgebra::BaseMatrix& W, const SReal regularization);
-
-    // Hot-start mechanism types
-    typedef core::behavior::BaseLagrangianConstraint::ConstraintBlockInfo ConstraintBlockInfo;
-    typedef core::behavior::BaseLagrangianConstraint::PersistentID PersistentID;
-    typedef core::behavior::BaseLagrangianConstraint::VecConstraintBlockInfo VecConstraintBlockInfo;
-    typedef core::behavior::BaseLagrangianConstraint::VecPersistentID VecPersistentID;
-
-    class ConstraintBlockBuf
-    {
-    public:
-        std::map<PersistentID, int> persistentToConstraintIdMap;
-        int nbLines{0}; ///< how many dofs (i.e. lines in the matrix) are used by each constraint
-    };
-
-    /// Compute initial guess for constraint forces from previous timestep
-    void computeInitialGuess();
-
-    /// Save constraint forces for use as initial guess in next timestep
-    void keepContactForcesValue();
-
-    /// Get constraint info (block info and persistent IDs) for hot-start
-    void getConstraintInfo(const core::ConstraintParams* cparams);
-
-    // Hot-start data storage
-    std::map<core::behavior::BaseLagrangianConstraint*, ConstraintBlockBuf> m_previousConstraints;
-    type::vector<SReal> m_previousForces;
-    VecConstraintBlockInfo m_constraintBlockInfo;
-    VecPersistentID m_constraintIds;
-    unsigned int m_numConstraints{0}; ///< Number of constraints from current/previous timestep
 
 private:
